third_party/WebKit/LayoutTests/css-parser/ascii-case-insensitive-matching.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
#a {
  width: 100px;
  height: 100px;
  background-color: green;
}

@support\017f (background-color: red) {
  #a {
    background-color: red;
  }
}
</style>
<div id=a></div>
<script>
test(function() {
  assert_equals(getComputedStyle(a).backgroundColor, "rgb(0, 128, 0)");
}, "Identifier values are matched ASCII case-insensitively");
</script>
